+++ package.json
@@ -11,15 +11,20 @@
     "@types/node": "^20.2.3",
     "@types/react": "^18.2.6",
     "@types/react-dom": "^18.2.4",
+    "@types/react-loadable": "^5.5.7",
+    "ahooks": "^3.7.7",
     "axios": "^1.4.0",
+    "clsx": "^1.2.1",
     "lodash": "^4.17.21",
     "lodash.debounce": "^4.0.8",
+    "or": "^0.2.0",
     "package-name": "^0.1.0",
     "path-browserify": "^1.0.1",
     "qs": "^6.11.1",
     "react": "^18.2.0",
     "react-content-loader": "^6.2.1",
     "react-dom": "^18.2.0",
+    "react-loadable": "^5.5.0",
     "react-paginate": "^8.2.0",
     "react-redux": "^8.0.5",
     "react-router": "^6.10.0",
