+++ src/App.tsx
@@ -1,13 +1,22 @@
 
+import React from 'react'
+import Loadable from 'react-loadable'
 import "./scss/app.scss";
 import Home from "./pages/Home";
-import NotFound from "./pages/NotFound";
 import { Route, Routes } from "react-router";
-import Cart from "./pages/Cart";
-
-
-import FullPizza from "./pages/FullPizza";
 import MainLayout from "./layouts/MainLayout";
+// const Cart = React.lazy(() => import(/* webpackChunkName:'Cart'  */'./pages/Cart'));
+
+const Cart = Loadable({
+  loader: () => import(/* webpackChunkName:'Cart'  */'./pages/Cart'),
+  loading:() => <div>Loading...</div>,
+})
+
+
+const FullPizza:React.FC = React.lazy(() => import(/* webpackChunkName:'FullPizza'  */'./pages/FullPizza').then(m => ({default:FullPizza})));
+const NotFound = React.lazy(() => import(/* webpackChunkName:'NotFound'  */'./pages/NotFound'));
+
+
 
 
 function App() {
@@ -15,9 +24,15 @@ function App() {
               <Routes>
                 <Route path="/" element={<MainLayout />}>
                   <Route path="" element={<Home />} />
-                  <Route path="cart" element={<Cart />} />
-                  <Route path="pizza/:id" element={<FullPizza />} />
-                  <Route path="*" element={<NotFound />} />
+                  <Route path="cart" element={<React.Suspense fallback={<div>Loading...</div>}>
+                    <Cart />
+                  </React.Suspense>} />
+                  <Route path="pizza/:id" element={<React.Suspense fallback={<div>Loading...</div>}>
+                    <FullPizza />
+                  </React.Suspense>} />
+                  <Route path="*" element={<React.Suspense fallback={<div>Loading...</div>}>
+                    <NotFound />
+                  </React.Suspense>} />
                   </Route>
               </Routes>
   );
