+++ src/components/Header.tsx
@@ -1,18 +1,27 @@
+import React from "react";
 import { Link, useLocation } from "react-router-dom"
-
 import logoSvg from "../assets/img/pizza-logo.svg"
-import Home from "../pages/Home";
 import Search from "./Search";
 import { useSelector } from "react-redux";
-import { selectCart } from "../redux/slices/cartSlice";
+import { selectCart } from "../redux/cart/selectors";
 
-function Header(){
+
+export const Header:React.FC = () =>{
   const {items, totalPrice} = useSelector(selectCart);
- const location = useLocation();
+  const location = useLocation();
+  const isMounted = React.useRef(false);
 
 
 
   const totalCount = items.reduce((sum: number, item: any) => sum + item.count, 0);
+
+  React.useEffect(() => {
+    if (isMounted.current){
+      const json = JSON.stringify(items);
+      localStorage.setItem('cart', json)
+    }
+    isMounted.current = true;
+  }, [items])
 
     return (
       <div className="header">
@@ -26,7 +35,7 @@ function Header(){
           </div>
         </div>
       </Link>
-      <Search />
+      {location.pathname !== '/cart' && <Search />}
       <div className="header__cart">
       {location.pathname !== '/cart' && (
           <Link to="/cart" className="button button--cart">
@@ -41,4 +50,3 @@ function Header(){
     ) 
   }
 
-  export default Header;
