+++ src/components/Pagination/index.tsx
@@ -9,7 +9,7 @@ type PaginationProps = {
 }
 
 
-const Pagination: React.FC<PaginationProps> = ( { currentPage, onChangePage } ) => {
+export const Pagination: React.FC<PaginationProps> = ( { currentPage, onChangePage } ) => {
     return (
     <ReactPaginate
     className={styles.root}
@@ -26,4 +26,3 @@ const Pagination: React.FC<PaginationProps> = ( { currentPage, onChangePage } ) 
 
 }
 
-export default Pagination;
