+++ src/components/PizzaBlock/Skeleton.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import ContentLoader from "react-content-loader"
 
-const Skeleton = () => (
+export const Skeleton = () => (
   <ContentLoader 
     className="pizza-block"
     speed={2}
@@ -19,4 +19,3 @@ const Skeleton = () => (
   </ContentLoader>
 )
 
-export default Skeleton
