+++ src/components/Search/index.tsx
@@ -2,8 +2,9 @@ import React from "react";
  //@ts-ignore
 import debounce from "lodash.debounce"
 import styles from "./Search.module.scss"
-import { setSearchValue } from "../../redux/slices/filterSlice";
+
 import { useDispatch } from "react-redux";
+import { setSearchValue } from "../../redux/filter/slice";
 
 
 export const Search: React.FC = () => {
