+++ src/components/index.ts
@@ -1,0 +1,11 @@
+export * from './PizzaBlock/Skeleton'
+export * from './PizzaBlock'
+export * from './Header'
+export * from './Categories'
+export * from './CartItem'
+export * from './CartEmpty'
+export * from './Search'
+export * from './Sort'
+export * from './Pagination'
+export * from './NotFoundBlock'
+
