+++ src/layouts/MainLayout.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Header from "../components/Header";
+import { Header } from "../components";
 import {  Outlet } from "react-router";
 
 const MainLayout: React.FC = () => {
