+++ src/pages/Cart.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import CartItem from "../components/CartItem";
-import { clearItem, selectCart } from "../redux/slices/cartSlice";
-import CartEmpty from "../components/CartEmpty";
+import { CartItem, CartEmpty } from "../components";
+import { selectCart } from "../redux/cart/selectors";
+import { clearItem } from "../redux/cart/slice";
 
 const Cart: React.FC = () => {
   const dispatch = useDispatch();
