+++ src/pages/FullPizza.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
+import { Link } from 'react-router-dom'
 import axios from 'axios';
 import { useParams, useNavigate } from 'react-router';
+import Home from './Home';
 
 
- const FullPizza: React.FC = () => {
+export const FullPizza: React.FC = () => {
   const [pizza, setPizza] = React.useState<{
     image: string;
     title: string;
@@ -34,7 +36,13 @@ import { useParams, useNavigate } from 'react-router';
         <img src={pizza.image}/>
         <h2>{pizza.title}</h2>
         <h4>{pizza.price} R</h4>
+    <Link to='/not'>
+    <button  className="button button--outline button--add">
+
+<span>Назад</span>
+</button>
+    </Link>
     </div>
   )
 }
-export default FullPizza;
+
