+++ src/redux/cart/slice.ts
@@ -1,28 +1,14 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { RootState } from '../store';
+import { CartSliceState } from './types';
+import { getCartFromLs } from '../../utils/getCartFromLs';
+import { CartItem } from './types';
+import { calcTotalPrice } from '../../utils/calcTotalPrice';
 
 
-export type CartItem = {
-    id: string,
-    title: string
-    price: number
-    image: string
-    type: string,
-    size: number,
-    count: number
-}
 
 
-interface CartSliceState {
-    totalPrice: number;
-    items: CartItem[];
-}
 
-
-const initialState: CartSliceState = {
-    totalPrice: 0,
-    items: [],
-};
+const initialState: CartSliceState = getCartFromLs();
 
 const cartSlice = createSlice({
   name: 'cart',
@@ -40,9 +26,7 @@ const cartSlice = createSlice({
             }); 
         }
 
-     state.totalPrice = state.items.reduce((sum, obj) => {
-        return obj.price * obj.count + sum;
-      }, 0)
+        state.totalPrice = calcTotalPrice(state.items)
     },
     minusItem(state, action){
         const findItem = state.items.find(obj => obj.id === action.payload);
@@ -61,8 +45,7 @@ const cartSlice = createSlice({
   },
 });
 
-export const selectCart = (state: RootState) => state.cart;
-export const selectCartItemById = (id: string) => (state:RootState) => state.cart.items.find((obj) => obj.id === id)
+
 
 
 export  const { addItem, removeItem, minusItem, clearItem,   } = cartSlice.actions;
