+++ src/redux/cart/types.ts
@@ -1,0 +1,15 @@
+export type CartItem = {
+    id: string,
+    title: string
+    price: number
+    image: string
+    type: string,
+    size: number,
+    count: number
+}
+
+
+export interface CartSliceState {
+    totalPrice: number;
+    items: CartItem[];
+}
