+++ src/redux/filter/selectors.ts
@@ -1,0 +1,6 @@
+import { RootState } from "../store";
+
+
+export const selectFilter = (state:RootState) => state.filter;
+export const selectSort = (state:RootState) => state.filter.sort;
+
