+++ src/redux/pizza/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from "../store";
+
+export const selectPizzaData = (state:RootState) => state.pizza;
