+++ src/utils/math.ts
@@ -1,0 +1,4 @@
+export function add(a:number, b:number) {
+  console.log(111);
+    return a + b;
+  }
