+++ tsconfig.json
@@ -17,5 +17,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": ["src", "src/@types"]
+  "include": ["src", "src/@types", "src/utils"]
 }
